requirements.txt
flask==3.0.0
opencv-python==4.9.0.80
face-recognition
numpy==1.26.4
pandas==2.2.2
pillow